Addressing linter issues
