+++ src/main/java/com/cse/helium/app/config/SwaggerConfig.java
@@ -1,7 +1,6 @@
 package com.cse.helium.app.config;
 
-import com.cse.helium.app.services.configuration.JSONConfigReader;
-import java.util.LinkedHashMap;
+import com.cse.helium.app.services.configuration.JsonConfigReader;
 import java.util.Map;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -9,19 +8,19 @@ import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @PropertySource(
-  value = "classpath:/static/swagger/helium.json",
-  factory = JSONConfigReader.class
+    value = "classpath:/static/swagger/helium.json",
+    factory = JsonConfigReader.class
 )
 @ConfigurationProperties
 public class SwaggerConfig {
 
-  private LinkedHashMap<String, String> info;
+  private Map<String, String> info;
 
   public Map<String, String> getInfo() {
     return this.info;
   }
 
   public void setInfo(Map<String, String> info) {
-    this.info = (LinkedHashMap<String, String>) info;
+    this.info = info;
   }
 }
